Migrated GUID to a separate crate.
