+++ crates/tool-guid/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "guid"
+version = "1.0.0"
+edition = "2021"
+authors = ["Breno RdV"]
+description = "Command-line tool for generating UUIDv4s with multiple modes, interval support, and clipboard integration for development workflows."
+repository = "https://github.com/brenordv/rusted-toolbox"
+
+[dependencies]
+shared = { path = "../shared" }
+clap = "4.5.45"
+anyhow = "1.0.99"
+ctrlc = "3.4.7"
+tracing = "0.1.41"
