+++ crates/tool-guid/src/main.rs
@@ -1,20 +1,20 @@
-use rusted_toolbox::tools::guid::cli_utils::{
-    get_cli_arguments, print_runtime_info, validate_cli_arguments,
-};
-use rusted_toolbox::tools::guid::guid_app::{
-    continuous_generation, copy_guid_to_clipboard, generate_once,
-};
-use shared::constants::general::{EXIT_CODE_INTERRUPTED_BY_USER, GUID_APP_NAME};
+use shared::constants::general::{EXIT_CODE_INTERRUPTED_BY_USER};
 use shared::logging::app_logger::LogLevel;
 use shared::logging::logging_helpers::initialize_log;
 use shared::system::tool_exit_helpers::{exit_error, exit_success, exit_with_code};
+use crate::cli_utils::{get_cli_arguments, print_runtime_info, validate_cli_arguments};
+use crate::guid_app::{continuous_generation, copy_guid_to_clipboard, generate_once};
 use tracing::error;
+
+mod cli_utils;
+mod models;
+mod guid_app;
 
 /// GUID generator tool.
 ///
 /// Parses arguments, validates configuration, and generates GUIDs either once or continuously.
 fn main() {
-    initialize_log(GUID_APP_NAME, LogLevel::Info);
+    initialize_log(env!("CARGO_PKG_NAME"), LogLevel::Info);
 
     let args = get_cli_arguments();
