+++ crates/tool-guid/src/models.rs
