--- crates/tools/src/tools/guid/mod.rs
@@ -1,3 +1,0 @@
-pub mod cli_utils;
-pub mod guid_app;
-pub mod models;
